backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-cas==1.6.0
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0
httpx==0.25.2
beautifulsoup4==4.12.2
pdfplumber==0.9.0
python-dotenv==1.0.0
numpy>=1.24.0
sentence-transformers==2.7.0
torch>=2.0.0
transformers>=4.36.0
